refect: ScoreCalculation Hand내부로 옮김
